Create .gitignore file to exclude folder "Soup Dumps" from repo.
I want to eventually port the simulator to nodejs and dump a bunch of data to a database.  I'll be able to get practice in database administration and statistics, and perhaps eventually find the exact mathematical limit of evolution.
In the meantime, I've saved the last generation of a few runs for 2 reasons:
1. As an example of what mutation and natural selection eventually lead to
2. In the hope that they'll eventually be useful